refactor: split Document delegate extensions
